+++ tests/core/grid_uniform_channel_with_meshed_cylinder.cc
@@ -70,17 +70,23 @@ main()
 
       run_test<2>("2D 2x1 channel, padded", "0,0:2,1:1,0.5:0.1:0.2:1:1:1:1");
 
-      run_test<2>("2D 2x1 channel, padded, refined once",
-                  "0,0:2,1:1,0.5:0.1:0.2:1:1:1:1",
-                  1);
-
-      run_test<3>("3D 2x1x1 channel, padded, boundary ids, and no TFI region",
+      run_test<2>("2D 2x1 channel, padded, refined once, meshed cylinder",
                   "0,0:2,1:1,0.5:0.1:0.2:1:1:1:1:1.0:2:false:true",
                   1);
 
-      run_test<3>("3D 2x1x1 channel, padded, boundary ids, and TFI region",
-                  "0,0:2,1:1,0.5:0.1:0.2:1:1:1:1:1.0:2:true:true",
+      run_test<3>(
+        "3D 2x1x1 channel, padded, boundary ids, meshed cylinder, and no TFI region",
+        "0,0:2,1:1,0.5:0.1:0.2:1:1:1:1:1.0:2:false:true:true",
+        1);
+
+      run_test<3>("3D 2x1x1 channel, padded, boundary ids, and no TFI region",
+                  "0,0:2,1:1,0.5:0.1:0.2:1:1:1:1:1.0:2:false:false:true",
                   1);
+
+      run_test<3>(
+        "3D 2x1x1 channel, padded, boundary ids, meshed cylinder, and TFI region",
+        "0,0:2,1:1,0.5:0.1:0.2:1:1:1:1:1.0:2:true:true:true",
+        1);
     }
   catch (std::exception &exc)
     {