+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to the Lethe project will be documented in this file.
 The changelog for the previous releases of Lethe are located in the release_notes folder.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Master] - 2026/06/01
+
+### Added
+
+- MINOR A new feature to the `grid_uniform_channel_with_meshed_cylinder` and `grid_uniform_channel_with_meshed_square_prism` is added so it is now possible to mesh or not the obstacle inside the channel. The default behavior of the meshes is also change so the obstacle is not meshed.[#1997](https://github.com/chaos-polymtl/lethe/pull/1997)
+
 ## [Master] - 2026/05/29
 
 ### Added